requirements.txt
django
django-admin-sortable2
django-froala-editor
djangorestframework
django-cors-headers
ipdb
